Create lib/canvas
